(EDIT) Fix formatting for terms
Prior to this change, the _Convert Specific Table of Excel Sheet to JSON_
post used inconsistent formatting for specific terms and did not format
all properties, variables, and specific values when needed.

Italics should only be used with underscores, like `_this_`, and only for
emphasis. This change replaced the italics for the sheet and table name
from `*Science* sheet` and `*Class 6*` to `**Science** sheet` and
`**Class 6**` and ensured the sheets and tables are always formatted.

This change also ensures the PowerShell variable names, which were written
with the `$` prefix, are wrapped in backticks as required by the style guide.

Finally, it also formatted the cell ID with backticks because it represents
a specific value and formatted the property key names as bold.

Using consistent formatting makes it easier for readers to understand what
is and is not referencing something specific to the code or implementation
and what is general context.