(EDIT) Remove heading and extra images
This change edits the _Convert Specific Table of Excel Sheet to JSON_
blog post to remove the H1 and extraneous images.

Blog posts never include an H1 because the post's title is the H1.

The first image, for the "normal" table, isn't interesting and its
purpose is fulfilled by the preceding text.

The last image of the JSON output is more accessibly captured by the
JSON in the codeblock.

This change also updates the alt text for the screenshot to be more
coherent for readers using a screen reader.